Set Docker ONBUILD triggers and build ARGs (#37)
* Modified the image to build a barebone version of the spark-hadoop-airflow toolchain.
This image is meant to be derived subsequently, where the project specific parameters
are then given.

Derived image will have to specify the spark user and group, provide the hadoop
configuration files, and the airflow DAGs files.

This modification is made to have a base image that supports different project users,
unlike previously where the user is already specified. The problem previously is that
when the user has to be changed, derivative images have to setup the necessary
environments and configurations in their dockerfile.

Instead, this modification encodes these steps in this image, but delay the actual
configuration till the image is derived. Thus, it would be easier to just specify
the correct configuration arguments in the derived images

* Update the formatting of the README.md file

* Update Dockerfile to only delay building the project specific stuff, e.g.
user, group, hadoop configurations, airflow dags

Added back requirements.txt and sample airflow dags for testing

* add hadoop conf dir as placeholders

* update README and ARGs for user and group name